+++ spec/acts_as_taggable_on/related_spec.rb
@@ -1,8 +1,12 @@
 require 'spec_helper'
 
 describe "Acts As Taggable On" do
-  before(:each) do
-    clean_database!
+  before :each do
+    DatabaseCleaner.start
+  end
+
+  after :each do
+    DatabaseCleaner.clean
   end
 
   describe "Related Objects" do
