+++ spec/acts_as_taggable_on/tags_helper_spec.rb
@@ -1,9 +1,8 @@
 require 'spec_helper'
 
 describe ActsAsTaggableOn::TagsHelper do
-  before(:each) do
-    clean_database!
-
+  before :each do
+    DatabaseCleaner.start
     @bob = TaggableModel.create(:name => "Bob Jones",  :language_list => "ruby, php")
     @tom = TaggableModel.create(:name => "Tom Marley", :language_list => "ruby, java")
     @eve = TaggableModel.create(:name => "Eve Nodd",   :language_list => "ruby, c++")
@@ -11,6 +10,10 @@ describe ActsAsTaggableOn::TagsHelper do
     @helper = class Helper
       include ActsAsTaggableOn::TagsHelper
     end.new
+  end
+
+  after :each do
+    DatabaseCleaner.clean
   end
 
   it "should yield the proper css classes" do
