+++ spec/acts_as_taggable_on/utils_spec.rb
@@ -2,11 +2,16 @@ require 'spec_helper'
 
 describe ActsAsTaggableOn::Utils do
   describe "like_operator" do
-    before(:each) do
-      clean_database!
+    before :each do
+      DatabaseCleaner.start
       TaggableModel.acts_as_taggable_on(:tags, :languages, :skills, :needs, :offerings)
       @taggable = TaggableModel.new(:name => "Bob Jones")
     end
+
+    after :each do
+      DatabaseCleaner.clean
+    end
+
 
     it "should return 'ILIKE' when the adapter is PostgreSQL" do
       TaggableModel.connection.stub(:adapter_name).and_return("PostgreSQL")
