+++ spec/database.yml.sample
@@ -1,4 +1,4 @@
-sqlite3:
+test:
   adapter: sqlite3
   database: acts_as_taggable_on.sqlite3
