+++ spec/internal/app/models/altered_inheriting_taggable_model.rb
@@ -1,0 +1,3 @@
+class AlteredInheritingTaggableModel < TaggableModel
+  acts_as_taggable_on :parts
+end
