+++ spec/internal/app/models/cached_model.rb
@@ -1,0 +1,3 @@
+class CachedModel < ActiveRecord::Base
+  acts_as_taggable
+end
