+++ spec/internal/app/models/inheriting_taggable_model.rb
@@ -1,0 +1,2 @@
+class InheritingTaggableModel < TaggableModel
+end
