+++ spec/internal/app/models/non_standard_id_taggable_model.rb
@@ -1,0 +1,8 @@
+class NonStandardIdTaggableModel < ActiveRecord::Base
+  primary_key = "an_id"
+  acts_as_taggable
+  acts_as_taggable_on :languages
+  acts_as_taggable_on :skills
+  acts_as_taggable_on :needs, :offerings
+  has_many :untaggable_models
+end
