+++ spec/internal/app/models/ordered_taggable_model.rb
@@ -1,0 +1,4 @@
+class OrderedTaggableModel < ActiveRecord::Base
+  acts_as_ordered_taggable
+  acts_as_ordered_taggable_on :colours
+end
