+++ spec/internal/app/models/other_cached_model.rb
@@ -1,0 +1,3 @@
+class OtherCachedModel < ActiveRecord::Base
+  acts_as_taggable_on :languages, :statuses, :glasses
+end
