+++ spec/internal/app/models/other_taggable_model.rb
@@ -1,0 +1,4 @@
+class OtherTaggableModel < ActiveRecord::Base
+  acts_as_taggable_on :tags, :languages
+  acts_as_taggable_on :needs, :offerings
+end
