+++ spec/internal/app/models/taggable_model.rb
@@ -1,0 +1,13 @@
+class TaggableModel < ActiveRecord::Base
+  acts_as_taggable
+  acts_as_taggable_on :languages
+  acts_as_taggable_on :skills
+  acts_as_taggable_on :needs, :offerings
+  has_many :untaggable_models
+
+  attr_reader :tag_list_submethod_called
+  def tag_list=v
+    @tag_list_submethod_called = true
+    super
+  end
+end
