+++ spec/internal/app/models/untaggable_model.rb
@@ -1,0 +1,3 @@
+class UntaggableModel < ActiveRecord::Base
+  belongs_to :taggable_model
+end
