+++ spec/internal/app/models/user.rb
@@ -1,0 +1,3 @@
+class User < ActiveRecord::Base
+  acts_as_tagger
+end
