+++ spec/internal/db/schema.rb
@@ -1,25 +1,4 @@
 ActiveRecord::Schema.define :version => 0 do
-  create_table :tags, :force => true do |t|
-    t.string :name
-  end
-
-  create_table :taggings, :force => true do |t|
-    t.references :tag
-
-    # You should make sure that the column created is
-    # long enough to store the required class names.
-    t.references :taggable, :polymorphic => true
-    t.references :tagger, :polymorphic => true
-
-    # Limit is created to prevent MySQL error on index
-    # length for MyISAM table type: http://bit.ly/vgW2Ql
-    t.string :context, :limit => 128
-
-    t.datetime :created_at
-  end
-  # above copied from
-  # generators/acts_as_taggable_on/migration/migration_generator
-
   create_table :taggable_models, :force => true do |t|
     t.column :name, :string
     t.column :type, :string
