+++ spec/internal/log/.gitignore
@@ -1,0 +1,1 @@
+*.log
